CKS: Support Ciilum and new dashboard
Usage:

    cd scripts/util/
    ./create-kubernetes-binaries-iso.sh ./ 1.34.1 1.8.0 1.34.0 \
    ./cks-config-cilium.yaml \
    ./cks-config-dashboard.yaml \
    cks-1.34.1-cilium-1.18.1

Refer to
https://docs.cloudstack.apache.org/en/latest/plugins/cloudstack-kubernetes-service.html#kubernetes-supported-versions

dashboard: https://artifacthub.io/packages/helm/k8s-dashboard/kubernetes-dashboard/7.13.0
cilium: https://artifacthub.io/packages/helm/cilium/cilium/1.18.1

+++ scripts/util/create-kubernetes-binaries-iso.sh
@@ -19,7 +19,7 @@
 set -e
 
 if [ $# -lt 6 ]; then
-    echo "Invalid input. Valid usage: ./create-kubernetes-binaries-iso.sh OUTPUT_PATH KUBERNETES_VERSION CNI_VERSION CRICTL_VERSION WEAVENET_NETWORK_YAML_CONFIG DASHBOARD_YAML_CONFIG BUILD_NAME [ARCH] [ETCD_VERSION]"
+    echo "Invalid input. Valid usage: ./create-kubernetes-binaries-iso.sh OUTPUT_PATH KUBERNETES_VERSION CNI_VERSION CRICTL_VERSION NETWORK_YAML_CONFIG DASHBOARD_YAML_CONFIG BUILD_NAME [ARCH] [ETCD_VERSION]"
     echo "eg: ./create-kubernetes-binaries-iso.sh ./ 1.11.4 0.7.1 1.11.1 https://github.com/weaveworks/weave/releases/download/latest_release/weave-daemonset-k8s-1.11.yaml https://raw.githubusercontent.com/kubernetes/dashboard/v1.10.0/src/deploy/recommended/kubernetes-dashboard.yaml setup-v1.11.4 amd64"
     exit 1
 fi
@@ -94,12 +94,20 @@ fi
 NETWORK_CONFIG_URL="${5}"
 echo "Downloading network config ${NETWORK_CONFIG_URL}"
 network_conf_file="${working_dir}/network.yaml"
-curl -sSL ${NETWORK_CONFIG_URL} -o ${network_conf_file}
+if [ -f "$NETWORK_CONFIG_URL" ];then
+  cp "${NETWORK_CONFIG_URL}" ${network_conf_file}
+else
+  curl -sSL ${NETWORK_CONFIG_URL} -o ${network_conf_file}
+fi
 
 DASHBORAD_CONFIG_URL="${6}"
 echo "Downloading dashboard config ${DASHBORAD_CONFIG_URL}"
 dashboard_conf_file="${working_dir}/dashboard.yaml"
-curl -sSL ${DASHBORAD_CONFIG_URL} -o ${dashboard_conf_file}
+if [ -f "$DASHBORAD_CONFIG_URL" ];then
+  cp "${DASHBORAD_CONFIG_URL}" ${dashboard_conf_file}
+else
+  curl -sSL ${DASHBORAD_CONFIG_URL} -o ${dashboard_conf_file}
+fi
 
 # TODO : Change the url once merged
 AUTOSCALER_URL="https://raw.githubusercontent.com/kubernetes/autoscaler/master/cluster-autoscaler/cloudprovider/cloudstack/examples/cluster-autoscaler-standard.yaml"
@@ -132,8 +140,15 @@ output=`${k8s_dir}/kubeadm config images list --kubernetes-version=${RELEASE}`
 # Don't forget about the yaml images !
 for i in ${network_conf_file} ${dashboard_conf_file}
 do
-  images=`grep "image:" $i | cut -d ':' -f2- | tr -d ' ' | tr -d "'"`
-  output=`printf "%s\n" ${output} ${images}`
+  if grep "^#cks-config" $i;then
+    images=$(python3 -c "import sys, yaml; [print(x) for x in yaml.safe_load(sys.stdin)['images']]" < $i)
+    for image in ${images}; do
+      output=`printf "%s\n" ${output} ${image}`
+    done
+  else
+    images=`grep "image:" $i | cut -d ':' -f2- | tr -d ' ' | tr -d "'"`
+    output=`printf "%s\n" ${output} ${images}`
+  fi
 done
 
 # Don't forget about the other image !
@@ -145,7 +160,7 @@ output=`printf "%s\n" ${output} ${provider_image}`
 
 while read -r line; do
     echo "Downloading image $line ---"
-    if [[ $line == kubernetesui* ]] || [[ $line == apache* ]] || [[ $line == weaveworks* ]]; then
+    if [[ $line == kubernetesui* ]] || [[ $line == apache* ]] || [[ $line == weaveworks* ]] || [[ $line == library* ]]; then
       line="docker.io/${line}"
     fi
     sudo ctr image pull "$line"
